refactor: 패키지 정리
- Command, Result, Web Request 관련 클래스를 dto 패키지로 정리
- 유틸리티 성격의 UidGenerator를 util 패키지로 정리
