+++ src/main/java/com/alzzaipo/hexagonal/ipo/adapter/in/web/IpoController.java
@@ -1,6 +1,8 @@
 package com.alzzaipo.hexagonal.ipo.adapter.in.web;
 
 import com.alzzaipo.hexagonal.ipo.application.port.in.*;
+import com.alzzaipo.hexagonal.ipo.application.port.in.dto.AnalyzeIpoProfitRateCommand;
+import com.alzzaipo.hexagonal.ipo.application.port.in.dto.AnalyzeIpoProfitRateResult;
 import com.alzzaipo.hexagonal.ipo.domain.Ipo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
