+++ src/main/java/com/alzzaipo/hexagonal/ipo/adapter/in/web/IpoScrapingController.java
@@ -1,7 +1,8 @@
 package com.alzzaipo.hexagonal.ipo.adapter.in.web;
 
+import com.alzzaipo.hexagonal.ipo.adapter.in.web.dto.ScrapeAndRegisterIposWebRequest;
 import com.alzzaipo.hexagonal.ipo.application.port.in.ScrapeAndRegisterIposUseCase;
-import com.alzzaipo.hexagonal.ipo.application.port.out.ScrapeIposCommand;
+import com.alzzaipo.hexagonal.ipo.application.port.out.dto.ScrapeIposCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
