+++ src/main/java/com/alzzaipo/hexagonal/ipo/application/port/in/AnalyzeIpoProfitRateQuery.java
@@ -1,5 +1,8 @@
 package com.alzzaipo.hexagonal.ipo.application.port.in;
 
+import com.alzzaipo.hexagonal.ipo.application.port.in.dto.AnalyzeIpoProfitRateCommand;
+import com.alzzaipo.hexagonal.ipo.application.port.in.dto.AnalyzeIpoProfitRateResult;
+
 public interface AnalyzeIpoProfitRateQuery {
 
     AnalyzeIpoProfitRateResult analyzeIpoProfitRate(AnalyzeIpoProfitRateCommand command);
