+++ src/main/java/com/alzzaipo/hexagonal/ipo/application/port/in/ScrapeAndRegisterIposUseCase.java
@@ -1,6 +1,6 @@
 package com.alzzaipo.hexagonal.ipo.application.port.in;
 
-import com.alzzaipo.hexagonal.ipo.application.port.out.ScrapeIposCommand;
+import com.alzzaipo.hexagonal.ipo.application.port.out.dto.ScrapeIposCommand;
 
 public interface ScrapeAndRegisterIposUseCase {
