+++ src/main/java/com/alzzaipo/hexagonal/ipo/application/port/in/dto/AnalyzeIpoProfitRateCommand.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.hexagonal.ipo.application.port.in;
+package com.alzzaipo.hexagonal.ipo.application.port.in.dto;
 
 import lombok.Getter;
