+++ src/main/java/com/alzzaipo/hexagonal/member/adapter/in/web/NewMemberController.java
@@ -3,7 +3,15 @@ package com.alzzaipo.hexagonal.member.adapter.in.web;
 import com.alzzaipo.hexagonal.common.Email;
 import com.alzzaipo.hexagonal.common.MemberPrincipal;
 import com.alzzaipo.hexagonal.common.Uid;
+import com.alzzaipo.hexagonal.member.adapter.in.web.dto.ChangeLocalAccountPasswordWebRequest;
+import com.alzzaipo.hexagonal.member.adapter.in.web.dto.LocalAccountPasswordDto;
+import com.alzzaipo.hexagonal.member.adapter.in.web.dto.LocalLoginWebRequest;
+import com.alzzaipo.hexagonal.member.adapter.in.web.dto.RegisterLocalAccountWebRequest;
 import com.alzzaipo.hexagonal.member.application.port.in.*;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.ChangeLocalAccountPasswordCommand;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.LocalLoginCommand;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.LocalLoginResult;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.RegisterLocalAccountCommand;
 import com.alzzaipo.hexagonal.member.domain.LocalAccount.LocalAccountId;
 import com.alzzaipo.hexagonal.member.domain.LocalAccount.LocalAccountPassword;
 import lombok.RequiredArgsConstructor;
