+++ src/main/java/com/alzzaipo/hexagonal/member/adapter/in/web/dto/LocalLoginWebRequest.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.hexagonal.member.adapter.in.web;
+package com.alzzaipo.hexagonal.member.adapter.in.web.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,10 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class RegisterLocalAccountWebRequest {
+public class LocalLoginWebRequest {
 
     private String accountId;
     private String accountPassword;
-    private String email;
-    private String nickname;
 }
