+++ src/main/java/com/alzzaipo/hexagonal/member/adapter/in/web/dto/RegisterLocalAccountWebRequest.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.hexagonal.member.adapter.in.web;
+package com.alzzaipo.hexagonal.member.adapter.in.web.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,8 +7,10 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class LocalLoginWebRequest {
+public class RegisterLocalAccountWebRequest {
 
     private String accountId;
     private String accountPassword;
+    private String email;
+    private String nickname;
 }
