+++ src/main/java/com/alzzaipo/hexagonal/member/application/port/in/ChangeLocalAccountPasswordUseCase.java
@@ -1,5 +1,7 @@
 package com.alzzaipo.hexagonal.member.application.port.in;
 
+import com.alzzaipo.hexagonal.member.application.port.in.dto.ChangeLocalAccountPasswordCommand;
+
 public interface ChangeLocalAccountPasswordUseCase {
 
     boolean changeLocalAccountPassword(ChangeLocalAccountPasswordCommand command);
