+++ src/main/java/com/alzzaipo/hexagonal/member/application/port/in/LocalLoginUseCase.java
@@ -1,5 +1,8 @@
 package com.alzzaipo.hexagonal.member.application.port.in;
 
+import com.alzzaipo.hexagonal.member.application.port.in.dto.LocalLoginCommand;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.LocalLoginResult;
+
 public interface LocalLoginUseCase {
 
     LocalLoginResult handleLocalLogin(LocalLoginCommand command);
