+++ src/main/java/com/alzzaipo/hexagonal/member/application/port/in/RegisterLocalAccountUseCase.java
@@ -1,5 +1,7 @@
 package com.alzzaipo.hexagonal.member.application.port.in;
 
+import com.alzzaipo.hexagonal.member.application.port.in.dto.RegisterLocalAccountCommand;
+
 public interface RegisterLocalAccountUseCase {
 
     void registerLocalAccount(RegisterLocalAccountCommand command);
