+++ src/main/java/com/alzzaipo/hexagonal/member/application/service/CheckLocalAccountPasswordService.java
@@ -3,7 +3,7 @@ package com.alzzaipo.hexagonal.member.application.service;
 import com.alzzaipo.hexagonal.common.Uid;
 import com.alzzaipo.hexagonal.member.application.port.in.CheckLocalAccountPasswordQuery;
 import com.alzzaipo.hexagonal.member.application.port.out.FindLocalAccountByMemberUidPort;
-import com.alzzaipo.hexagonal.member.application.port.out.SecureLocalAccount;
+import com.alzzaipo.hexagonal.member.application.port.out.dto.SecureLocalAccount;
 import com.alzzaipo.hexagonal.member.domain.LocalAccount.LocalAccountPassword;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
