+++ src/main/java/com/alzzaipo/hexagonal/member/application/service/RegisterLocalAccountService.java
@@ -5,11 +5,11 @@ import com.alzzaipo.hexagonal.email.application.port.in.CheckEmailVerifiedQuery;
 import com.alzzaipo.hexagonal.email.application.port.out.DeleteOldEmailVerificationHistoryPort;
 import com.alzzaipo.hexagonal.member.application.port.in.CheckLocalAccountEmailAvailabilityQuery;
 import com.alzzaipo.hexagonal.member.application.port.in.CheckLocalAccountIdAvailabilityQuery;
-import com.alzzaipo.hexagonal.member.application.port.in.RegisterLocalAccountCommand;
+import com.alzzaipo.hexagonal.member.application.port.in.dto.RegisterLocalAccountCommand;
 import com.alzzaipo.hexagonal.member.application.port.in.RegisterLocalAccountUseCase;
 import com.alzzaipo.hexagonal.member.application.port.out.RegisterLocalAccountPort;
 import com.alzzaipo.hexagonal.member.application.port.out.RegisterMemberPort;
-import com.alzzaipo.hexagonal.member.application.port.out.SecureLocalAccount;
+import com.alzzaipo.hexagonal.member.application.port.out.dto.SecureLocalAccount;
 import com.alzzaipo.hexagonal.member.domain.LocalAccount.LocalAccountId;
 import com.alzzaipo.hexagonal.member.domain.Member.Member;
 import lombok.RequiredArgsConstructor;
